Break module into files
