+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/components.rs
@@ -1,6 +1,7 @@
 pub mod actions;
 pub mod common;
 pub mod component_creation;
+pub mod create_components;
 pub mod doenet;
 pub mod prelude;
 pub mod special;
