+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/dast.rs
@@ -1,8 +1,8 @@
 pub mod dast_conversion;
 pub mod dast_structure;
+pub mod flat_dast;
 pub mod macro_expand;
 pub mod macro_resolve;
-pub mod untagged_flat_dast;
 
 pub use dast_conversion::*;
 pub use dast_structure::*;
